backend/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
groq
requests
python-dotenv
beautifulsoup4
lxml
numpy
langdetect
